Remove session timezone setting from the generic_sql query-processor [ci drivers]
This can potentially cause issues as JDBC's behavior is to convert a
timestamp to the JVM's default time zone. If this value is set to
something other than the JVM's timezone, incorrect timezone
conversions will be done and users will get incorrect results. This
conversion is now happening as Date/Timestamp objects are being
constructed and use the detected timezone of the database.